Adding Undo/Redo && updating LocalStorage

+++ src/app/todo-list/todo-list.component.ts
@@ -1,6 +1,7 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, HostListener, OnInit} from '@angular/core';
 import {TodoListData} from '../dataTypes/TodoListData';
 import {TodoService} from '../services/todo.service';
+import {Hotkeys} from '../services/hotkeys.service';
 
 @Component({
     selector: 'app-todo-list',
@@ -12,11 +13,17 @@ export class TodoListComponent implements OnInit {
 
     private todoList: TodoListData;
 
-    constructor(private todoService: TodoService) {
+    constructor(private todoService: TodoService, private hotkeysService: Hotkeys) {
         todoService.getTodoListDataObservable().subscribe( tdl => this.todoList = tdl );
     }
 
+    // initialis les hotkeys
     ngOnInit() {
+        // Undo
+        // this.hotkeysService.addShortcut({ keys: 'control.z' }).subscribe(x => this.undoTodolist(this.todoService));
+
+        // Redo
+        // this.hotkeys.addShortcut({ keys: 'control.shift.z' }).subscribe(this.todoService.redoTodolist);
     }
 
     get label(): string {
@@ -24,7 +31,10 @@ export class TodoListComponent implements OnInit {
     }
 
     get todoCount(): number {
-        return this.todoList.items.filter(I => I.isDone === false && I.isDeleted === false).length;
+        if (this.todoList.items) {
+            return this.todoList.items.filter(I => I.isDone === false && I.isDeleted === false).length;
+        }
+        return 0;
     }
 
     itemShow(type: string) {
@@ -32,8 +42,10 @@ export class TodoListComponent implements OnInit {
     }
 
     setDoneAllItems() {
-        this.todoService.todoListSettingAllItemStatus = !this.todoService.todoListSettingAllItemStatus;
-        this.todoList.items.map(item => this.todoService.setItemsDone(this.todoService.todoListSettingAllItemStatus, item));
+        if (this.todoList.items) {
+            this.todoService.todoListSettingAllItemStatus = !this.todoService.todoListSettingAllItemStatus;
+            this.todoList.items.map(item => this.todoService.setItemsDone(this.todoService.todoListSettingAllItemStatus, item));
+        }
     }
 
     appendItem(input: any) {
@@ -58,10 +70,25 @@ export class TodoListComponent implements OnInit {
     }
 
     itemDoneRemove() {
-        this.todoList.items.filter(item => item.isDone === true && this.todoService.removeItems(item));
+        console.log(this.todoService.todoListExistItemDone);
+        if (this.todoList.items) {
+            this.todoList.items.filter(item => item.isDone === true && this.todoService.removeItems(item));
+        }
     }
 
     itemAllRemove() {
-        this.todoService.removeAllItems();
+        if (this.todoList.items) {
+            this.todoService.removeAllItems();
+        }
+    }
+
+    @HostListener('document:keydown.control.z')
+    undoTodolist() {
+        this.todoService.restoreTodoList('undo');
+    }
+
+    @HostListener('document:keydown.control.shift.z')
+    redoTodolist() {
+        this.todoService.restoreTodoList('redo');
     }
 }
